+++ R/cffr-package.R
@@ -8,5 +8,5 @@ NULL
 # import stuffs
 #' @importFrom utils download.file as.person read.csv person capture.output
 #' installed.packages packageDescription citation available.packages
-#' modifyList toBibtex bibentry
+#' modifyList toBibtex bibentry head tail
 NULL